fix(auth): show change-password and logout in the sidebar on every page
