+++ frontend/src/layouts/RootLayout.vue
@@ -1,35 +1,18 @@
 <script setup>
-import { ref } from 'vue'
 import AppSidebar from '../components/layout/AppSidebar.vue'
-import { useRouter } from 'vue-router'
-import AppButton from '../components/ui/AppButton.vue'
-import ChangePasswordModal from '../components/modals/ChangePasswordModal.vue'
-import { useAuthStore } from '../stores/auth'
 
 // `showCreateModal` is provided at the App.vue level so the create-server
 // modal works from any layout/route (e.g. ServerSwitcher's "Add server").
-
-const auth = useAuthStore()
-const router = useRouter()
-const showChangePassword = ref(false)
-
-async function onLogout() {
-  await auth.logout()
-  router.push({ name: 'Login' })
-}
+// Account actions (change password / log out) live in AppSidebar so they
+// are reachable on every authenticated page, including server pages.
 </script>
 
 <template>
   <div class="root-layout">
     <AppSidebar />
     <main class="root-layout__content">
-      <div v-if="auth.enabled" class="root-layout__topbar">
-        <AppButton variant="ghost" size="sm" @click="showChangePassword = true">Change password</AppButton>
-        <AppButton variant="ghost" size="sm" @click="onLogout">Log out</AppButton>
-      </div>
       <router-view />
     </main>
-    <ChangePasswordModal :show="showChangePassword" @close="showChangePassword = false" />
   </div>
 </template>
 
@@ -47,12 +30,5 @@ async function onLogout() {
   flex-direction: column;
   min-width: 0;
   overflow-y: auto;
-}
-
-.root-layout__topbar {
-  display: flex;
-  justify-content: flex-end;
-  gap: var(--space-2);
-  padding: var(--space-3) var(--space-4);
 }
 </style>
